requirements.txt
python-dotenv
google-cloud-vision
opencv-python
Pillow
PyMuPDF
pandas
xlsxwriter